fix: Improve PDF stamping robustness and safety
Adds checks for null/undefined invoice amounts before attempting to stamp.
Ensures `stampedPdfBytes` is correctly cast to `Uint8Array` when creating the Blob for PDF generation.
Includes a null check for the Supabase client in `App.tsx` to prevent errors during initialization.
Safely handles potentially undefined user email by coalescing to `null` when setting `setupAccountEmail`.

+++ components/InvoiceDetail.tsx
@@ -196,7 +196,7 @@ const InvoiceDetail: React.FC<InvoiceDetailProps> = ({ invoice, fileData, projec
           const stampedPdfBytes = await stampInvoicePdf(loadedFileContent, invoice, project, allocations);
           
           // Trigger Download
-          const blob = new Blob([stampedPdfBytes], { type: 'application/pdf' });
+          const blob = new Blob([stampedPdfBytes as Uint8Array], { type: 'application/pdf' }); // Cast to Uint8Array
           const url = window.URL.createObjectURL(blob);
           const link = document.createElement('a');
           link.href = url;
